… examples/stable-diffusion/training/README.md