# What does this PR do?

Speeds up FLUX training **more than 2x** by using **Gaudi optimized attention**.

FLUX train speed **before fix** (1xG2 w/ 1.20.0-543 release) = 5.19s/it
```bash
05/01/2025 17:29:03 - INFO - accelerate.checkpointing - Optimizer state saved in dog_lora_flux/checkpoint-500/optimizer.bin
05/01/2025 17:29:03 - INFO - accelerate.checkpointing - Scheduler state saved in dog_lora_flux/checkpoint-500/scheduler.bin
05/01/2025 17:29:03 - INFO - accelerate.checkpointing - Sampler state for dataloader 0 saved in dog_lora_flux/checkpoint-500/sampler.bin
05/01/2025 17:29:03 - INFO - accelerate.checkpointing - Random states saved in dog_lora_flux/checkpoint-500/random_states_0.pkl
05/01/2025 17:29:03 - INFO - main - Saved state to dog_lora_flux/checkpoint-500
Steps: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 500/500 
[54:01<00:00,  5.19s/it, loss=0.393, lr=0.0001]
```

FLUX train speed **AFTER fix** (1xG2 w/ 1.20.0-543 release) = 2.49s/it
```bash
05/01/2025 16:21:30 - INFO - accelerate.checkpointing - Optimizer state saved in dog_lora_flux/checkpoint-500/optimizer.bin
05/01/2025 16:21:30 - INFO - accelerate.checkpointing - Scheduler state saved in dog_lora_flux/checkpoint-500/scheduler.bin
05/01/2025 16:21:30 - INFO - accelerate.checkpointing - Sampler state for dataloader 0 saved in dog_lora_flux/checkpoint-500/sampler.bin
05/01/2025 16:21:30 - INFO - accelerate.checkpointing - Random states saved in dog_lora_flux/checkpoint-500/random_states_0.pkl
05/01/2025 16:21:30 - INFO - __main__ - Saved state to dog_lora_flux/checkpoint-500
Steps: 100%|███████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████████| 500/500 
[27:28<00:00,  2.49s/it, loss=0.395, lr=0.0001]
```

FLUX Multi-card training is also improved similarly (for default DP, MPI and DeepSpeed cases).