@emascarenhas 2025-05-14
Please do a "make style" to ensure no issues.
Also, I suppose you've run the pytests for stable-diffusion, please confirm.