@dsocek 2025-05-15
@emascarenhas thanks for review.  I ran `make style` before submitting (just confirmed), and also I used sandbox to run CI tests for diffusers which pass all diffuser tests.
![image](https://github.com/user-attachments/assets/36c0a2c9-1294-4441-aac5-50834f4984a0)