@emascarenhas 2025-05-15
> @emascarenhas thanks for review. I ran `make style` before submitting (just confirmed), and also I used sandbox to run CI tests for diffusers which pass all diffuser tests. ![image](https://private-user-images.githubusercontent.com/62973902/444179735-36c0a2c9-1294-4441-aac5-50834f4984a0.png?
Thanks. Can you respond to the other comments above.