@HuggingFaceDocBuilderDev 2025-05-19
The docs for this PR live [here](https://moon-ci-docs.huggingface.co/docs/optimum-habana/pr_1963). All of your documentation changes will be reflected on that endpoint. The docs are available until 30 days after the last update.